requirements-windows.txt
PyGObject
ytmusicapi
yt-dlp
yt-dlp-ejs
requests
urllib3
mutagen
StrEnum==0.4.15
pywebview
pystray
Pillow
